refactor(ts): update WorldData to Uint16Array and use tile level in renderer
- WorldData now reads packed u16 tiles (type 4bit<<12 | level 4bit<<8 | variant 2bit<<6)
- Added getTileLevel() method to extract level from packed tile data
- GamePage uses Uint16Array for WASM tile memory view
- IsometricCanvas passes actual tile level from WorldData to drawTile instead of hardcoded 8
- Rebuilt WASM with wasm-pack for web target

+++ src/pages/game/ui/GamePage.tsx
@@ -42,7 +42,7 @@ export const GamePage: React.FC = () => {
 		const d = world_depth();
 		const h = world_height();
 
-		const tiles = new Uint8Array(wasmOutput.memory.buffer, ptr, len);
+		const tiles = new Uint16Array(wasmOutput.memory.buffer, ptr, len);
 		return new WorldData(w, d, h, tiles);
 	}, [isSuccess, wasmOutput]);
